% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lerp.R
\name{lerp}
\alias{lerp}
\title{Mixbox interpolation between two colors}
\usage{
lerp(x, y, t)
}
\arguments{
\item{x, y}{Colors to interpolate between}

\item{t}{Mixing ratio}
}
\value{
Returns interpolated colors as native packed integers.
For \code{lerp.nativeRaster()}, it returns a \code{nativeRaster} object.
}
\description{
Blends two colors using the 'Mixbox' algorithm.
The input is expected to be an integer vector with the color
channels packed in the following way:
\itemize{
\item red: bits 0-7
\item green: bits 8-15
\item blue: bits 16-23
\item alpha: bits 24-31
}

You can use \code{\link[colorfast:col_to_int]{colorfast::col_to_int()}} to convert colors to integers
and \code{\link[colorfast:int_to_col]{colorfast::int_to_col()}} to convert integers back to hexdecimal colors.
}
\seealso{
\itemize{
\item \href{https://scrtwpns.com/mixbox/}{Mixbox - Natural Color Mixing Based on Real Pigments}
\item \href{https://github.com/scrtwpns/mixbox}{scrtwpns/mixbox}
}
}
