% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixboxr-package.R
\docType{package}
\name{mixboxr-package}
\alias{mixboxr}
\alias{mixboxr-package}
\title{mixboxr: Color Blending with 'Mixbox'}
\description{
Offers blending method for natural color mixing with the C/C++ implmentation of 'Mixbox' \url{https://github.com/scrtwpns/mixbox} library.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/paithiov909/mixboxr/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Akiru Kato \email{paithiov909@gmail.com}
}

Other contributors:
\itemize{
  \item Secret Weapons \email{mixbox@scrtwpns.com} [copyright holder]
}

}
\keyword{internal}
