% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lerp.R
\name{lerp}
\alias{lerp}
\title{Mixbox interpolation between two colors}
\usage{
lerp(x, y, t)
}
\arguments{
\item{x, y}{Colors to interpolate between}

\item{t}{Mixing ratio}
}
\value{
Interpolated color
}
\description{
Blends two colors using the 'mixbox' algorithm.
The input is expected to be an integer vector with the color
channels packed in the following way:
\itemize{
\item red: bits 0-7
\item green: bits 8-15
\item blue: bits 16-23
\item alpha: bits 24-31
}

You can use \code{\link[colorfast:col_to_int]{colorfast::col_to_int()}} to convert colors to integers
and \code{\link[colorfast:int_to_col]{colorfast::int_to_col()}} to convert back integers to hexdecimal colors.
}
